clean docs - udpate css - create baseline folder
