+++ docs/api_documentation/core/cma_mega.md
@@ -1,0 +1,5 @@
+# Covariance Matrix Adaptation MAP-Elites via Gradient Arborescence (CMA-MEGA)
+
+To create an instance of CMA-MEGA, one need to use an instance of [MAP-Elites](map_elites.md) with the CMAMEGAEmitter, detailed below.
+
+::: qdax.core.emitters.cma_mega_emitter.CMAMEGAEmitter
