+++ docs/api_documentation/core/omg_mega.md
@@ -1,0 +1,5 @@
+# Objective and Measure Gradient MAP-Elites via Gradient Arborescence (OMGMEGA)
+
+To create an instance of OMGMEGA, one need to use an instance of [MAP-Elites](map_elites.md) with the OMGMEGAEmitter, detailed below.
+
+::: qdax.core.emitters.omg_mega_emitter.OMGMEGAEmitter
