+++ docs/api_documentation/core/pgame.md
@@ -1,0 +1,5 @@
+# Policy Gradient Assisted MAP Elites (PGAME)
+
+To create an instance of PGAME, one need to use an instance of [MAP-Elites](map_elites.md) with the PGAMEEmitter, detailed below.
+
+::: qdax.core.emitters.pga_me_emitter.PGAMEEmitter
