+++ mkdocs.yml
@@ -9,7 +9,7 @@ strict: false
 
 
 theme:
-  name: material
+  name: readthedocs
   language: en
   palette:
     # Light mode / dark mode
@@ -27,10 +27,7 @@ theme:
       toggle:
         icon: material/weather-sunny
         name: Switch to light mode
-  #logo: img/parl-logo.png
-  # favicon: img/parl-logo.png
-  # icon:
-  #   repo: fontawesome/brands/gitlab
+
   features:
     - navigation.tracking
     - navigation.tabs
@@ -79,25 +76,29 @@ markdown_extensions:
   - pymdownx.tabbed:
      alternate_style: true
 
+extra_css:
+  - css/mkdocstrings.css
 
 plugins:
-  - search
-  - autorefs  # Cross-links to headings
-  - git-revision-date
-  - mknotebooks  # enable Jupyter notebooks
-  - mkdocstrings:
-      handlers:
-        python:
-          docstring_style: google
-          selection:
-            inherited_members: true  # Allow looking up inherited methods
-            filters:
-              - '!^test'
-              - "!^_"  # exclude all members starting with _
-              - "^__init__$"  # but always include __init__ modules and methods
-          rendering:
-            show_source: false # don't include source code (too cumbersome)
-            members_order: source  # order methods according to their order of definition in the source code, not alphabetical order
+- search
+- autorefs  # Cross-links to headings
+- git-revision-date
+- mknotebooks  # enable Jupyter notebooks
+- mkdocstrings:
+    custom_templates: templates
+    handlers:
+      python:
+        docstring_style: google
+        selection:
+          inherited_members: true  # Allow looking up inherited methods
+          filters:
+            - '!^test'
+            - "!^_"  # exclude all members starting with _
+            - "^__init__$"  # but always include __init__ modules and methods
+        rendering:
+          show_source: true # don't include source code (too cumbersome)
+          show_root_heading: true
+          members_order: source  # order methods according to their order of definition in the source code, not alphabetical order
 
 
 nav:
@@ -116,17 +117,22 @@ nav:
     - SMERL: notebooks/smerl_example.ipynb
   - API documentation:
     - Core:
-      - MAP Elites: api_documentation/core/map_elites.md
-      - MOME: api_documentation/core/mome.md
-      - SMERL: api_documentation/core/smerl.md
-      - DIAYN: api_documentation/core/diayn.md
-      - DADS: api_documentation/core/dads.md
-      - SAC: api_documentation/core/sac.md
-      - TD3: api_documentation/core/td3.md
-      - CMAES: api_documentation/core/cmaes.md
-      - Genetic Algorithm: api_documentation/core/genetic_algorithm.md
-      - NSGA2: api_documentation/core/nsga2.md
-      - SPEA2: api_documentation/core/spea2.md
+      - Core algorithms:
+        - MAP Elites: api_documentation/core/map_elites.md
+        - PGAME: api_documentation/core/pgame.md
+        - OMG MEGA: api_documentation/core/omg_mega.md
+        - CMA MEGA: api_documentation/core/cma_mega.md
+        - MOME: api_documentation/core/mome.md
+      - Baseline algorithms:
+        - SMERL: api_documentation/core/smerl.md
+        - DIAYN: api_documentation/core/diayn.md
+        - DADS: api_documentation/core/dads.md
+        - SAC: api_documentation/core/sac.md
+        - TD3: api_documentation/core/td3.md
+        - CMAES: api_documentation/core/cmaes.md
+        - Genetic Algorithm: api_documentation/core/genetic_algorithm.md
+        - NSGA2: api_documentation/core/nsga2.md
+        - SPEA2: api_documentation/core/spea2.md
       - Containers: api_documentation/core/containers.md
       - Emitters: api_documentation/core/emitters.md
       - Neuroevolution: api_documentation/core/neuroevolution.md
