+++ qdax/baselines/diayn.py
@@ -1,6 +1,6 @@
 """
-A collection of functions and classes that define the algorithm Dynamics Aware Discovery
-of Skills (DADS), ref: https://arxiv.org/abs/1907.01657.
+A collection of functions and classes that define the algorithm Diversity Is All You
+Need (DIAYN), ref: https://arxiv.org/abs/1802.06070.
 """
 
 from dataclasses import dataclass
@@ -14,22 +14,17 @@ from brax.envs import Env
 from brax.envs import State as EnvState
 from brax.training.distribution import NormalTanhDistribution
 
+from qdax.baselines.sac import SAC, SacConfig
 from qdax.core.neuroevolution.buffers.buffer import QDTransition, ReplayBuffer
-from qdax.core.neuroevolution.losses.dads_loss import make_dads_loss_fn
+from qdax.core.neuroevolution.losses.diayn_loss import make_diayn_loss_fn
 from qdax.core.neuroevolution.mdp_utils import TrainingState, get_first_episode
-from qdax.core.neuroevolution.networks.dads_networks import make_dads_networks
-from qdax.core.neuroevolution.normalization_utils import (
-    RunningMeanStdState,
-    normalize_with_rmstd,
-    update_running_mean_std,
-)
+from qdax.core.neuroevolution.networks.diayn_networks import make_diayn_networks
 from qdax.core.neuroevolution.sac_utils import generate_unroll
-from qdax.core.sac import SAC, SacConfig
 from qdax.types import Metrics, Params, Reward, RNGKey, Skill, StateDescriptor
 
 
-class DadsTrainingState(TrainingState):
-    """Contains training state for the learner."""
+class DiaynTrainingState(TrainingState):
+    """Training state for the DIAYN algorithm"""
 
     policy_optimizer_state: optax.OptState
     policy_params: Params
@@ -38,24 +33,22 @@ class DadsTrainingState(TrainingState):
     alpha_optimizer_state: optax.OptState
     alpha_params: Params
     target_critic_params: Params
-    dynamics_optimizer_state: optax.OptState
-    dynamics_params: Params
+    discriminator_optimizer_state: optax.OptState
+    discriminator_params: Params
     random_key: RNGKey
     steps: jnp.ndarray
-    normalization_running_stats: RunningMeanStdState
 
 
 @dataclass
-class DadsConfig(SacConfig):
+class DiaynConfig(SacConfig):
+    """Configuration for the DIAYN algorithm"""
+
     num_skills: int = 5
-    dynamics_update_freq: int = 1
     descriptor_full_state: bool = False
-    normalize_target: bool = True
-    omit_input_dynamics_dim: int = 2
 
 
-class DADS(SAC):
-    """Implements DADS algorithm https://arxiv.org/abs/1907.01657.
+class DIAYN(SAC):
+    """Implements DIAYN algorithm https://arxiv.org/abs/1802.06070.
 
     Note that the functions select_action, _update_alpha, _update_critic and
     _update_actor are inherited from SAC algorithm.
@@ -68,22 +61,24 @@ class DADS(SAC):
     of skills, is used to evaluate the skills in the environment and hence
     to generate transitions. The sampling is hence fixed and perfectly uniform.
 
+    Since we are using categorical skills, the current loss function used
+    to train the discriminator is the categorical cross entropy loss.
+
     We plan to add continous skill as an option in the future. We also plan
     to release the current constraint on the number of batched environments
     by sampling from the skills rather than having this fixed setting.
     """
 
-    def __init__(self, config: DadsConfig, action_size: int, descriptor_size: int):
-        self._config: DadsConfig = config
-
+    def __init__(self, config: DiaynConfig, action_size: int):
+        self._config: DiaynConfig = config
         if self._config.normalize_observations:
-            raise NotImplementedError("Normalization in not implemented for DADS yet")
+            raise NotImplementedError("Normalization is not implemented for DIAYN yet")
 
         # define the networks
-        self._policy, self._critic, self._dynamics = make_dads_networks(
+        self._policy, self._critic, self._discriminator = make_diayn_networks(
+            num_skills=self._config.num_skills,
             action_size=action_size,
-            descriptor_size=descriptor_size,
-            omit_input_dynamics_dim=config.omit_input_dynamics_dim,
+            hidden_layer_sizes=self._config.hidden_layer_sizes,
         )
 
         # define the action distribution
@@ -95,11 +90,11 @@ class DADS(SAC):
             self._alpha_loss_fn,
             self._policy_loss_fn,
             self._critic_loss_fn,
-            self._dynamics_loss_fn,
-        ) = make_dads_loss_fn(
+            self._discriminator_loss_fn,
+        ) = make_diayn_loss_fn(
             policy_fn=self._policy.apply,
             critic_fn=self._critic.apply,
-            dynamics_fn=self._dynamics.apply,
+            discriminator_fn=self._discriminator.apply,
             reward_scaling=self._config.reward_scaling,
             discount=self._config.discount,
             action_size=action_size,
@@ -111,7 +106,9 @@ class DADS(SAC):
         self._policy_optimizer = optax.adam(learning_rate=self._config.learning_rate)
         self._critic_optimizer = optax.adam(learning_rate=self._config.learning_rate)
         self._alpha_optimizer = optax.adam(learning_rate=self._config.learning_rate)
-        self._dynamics_optimizer = optax.adam(learning_rate=self._config.learning_rate)
+        self._discriminator_optimizer = optax.adam(
+            learning_rate=self._config.learning_rate
+        )
 
     def init(  # type: ignore
         self,
@@ -119,7 +116,7 @@ class DADS(SAC):
         action_size: int,
         observation_size: int,
         descriptor_size: int,
-    ) -> DadsTrainingState:
+    ) -> DiaynTrainingState:
         """Initialise the training state of the algorithm.
 
         Args:
@@ -127,16 +124,16 @@ class DADS(SAC):
             action_size: the size of the environment's action space
             observation_size: the size of the environment's observation space
             descriptor_size: the size of the environment's descriptor space (i.e. the
-                dimension of the dynamics network's input)
+                dimension of the discriminator's input)
 
         Returns:
-            the initial training state of DADS
+            the initial training state of DIAYN
         """
-        # Initialize params
+
+        # define policy and critic params
         dummy_obs = jnp.zeros((1, observation_size + self._config.num_skills))
         dummy_action = jnp.zeros((1, action_size))
-        dummy_dyn_obs = jnp.zeros((1, descriptor_size))
-        dummy_skill = jnp.zeros((1, self._config.num_skills))
+        dummy_discriminator_obs = jnp.zeros((1, descriptor_size))
 
         random_key, subkey = jax.random.split(random_key)
         policy_params = self._policy.init(subkey, dummy_obs)
@@ -149,21 +146,20 @@ class DADS(SAC):
         )
 
         random_key, subkey = jax.random.split(random_key)
-        dynamics_params = self._dynamics.init(
-            subkey,
-            obs=dummy_dyn_obs,
-            skill=dummy_skill,
-            target=dummy_dyn_obs,
+        discriminator_params = self._discriminator.init(
+            subkey, obs=dummy_discriminator_obs
         )
 
         policy_optimizer_state = self._policy_optimizer.init(policy_params)
         critic_optimizer_state = self._critic_optimizer.init(critic_params)
-        dynamics_optimizer_state = self._dynamics_optimizer.init(dynamics_params)
+        discriminator_optimizer_state = self._discriminator_optimizer.init(
+            discriminator_params
+        )
 
         log_alpha = jnp.asarray(jnp.log(self._config.alpha_init), dtype=jnp.float32)
         alpha_optimizer_state = self._alpha_optimizer.init(log_alpha)
 
-        return DadsTrainingState(
+        return DiaynTrainingState(
             policy_optimizer_state=policy_optimizer_state,
             policy_params=policy_params,
             critic_optimizer_state=critic_optimizer_state,
@@ -171,84 +167,56 @@ class DADS(SAC):
             alpha_optimizer_state=alpha_optimizer_state,
             alpha_params=log_alpha,
             target_critic_params=target_critic_params,
-            dynamics_optimizer_state=dynamics_optimizer_state,
-            dynamics_params=dynamics_params,
+            discriminator_optimizer_state=discriminator_optimizer_state,
+            discriminator_params=discriminator_params,
             random_key=random_key,
-            normalization_running_stats=RunningMeanStdState(
-                mean=jnp.zeros(
-                    descriptor_size,
-                ),
-                var=jnp.ones(
-                    descriptor_size,
-                ),
-                count=jnp.zeros(()),
-            ),
             steps=jnp.array(0),
         )
 
-    @partial(jax.jit, static_argnames=("self",))
+    @partial(jax.jit, static_argnames=("self", "add_log_p_z"))
     def _compute_diversity_reward(
-        self, transition: QDTransition, training_state: DadsTrainingState
+        self,
+        transition: QDTransition,
+        discriminator_params: Params,
+        add_log_p_z: bool = True,
     ) -> Reward:
-        """Computes the diversity reward of DADS.
+        """Computes the diversity reward of DIAYN.
+
+        As we use discrete skills (one hot encoded vectors), the categorical
+        cross entropy is used here.
 
         Args:
             transition: a batch of transitions from the replay buffer
-            training_state: the current training state
+            discriminator_params: the parameters of the discriminator
+            add_log_p_z: whether or not to add (minus) the probability of the skills'
+                prior distribution. Defaults to True.
 
         Returns:
             the diversity reward
         """
-        active_skills = transition.obs[:, -self._config.num_skills :]
 
-        # Compute dynamics prob
         next_state_desc = transition.next_state_desc
-        state_desc = transition.state_desc
-        target = next_state_desc - state_desc
-
-        if self._config.normalize_target:
-            target = normalize_with_rmstd(
-                target, training_state.normalization_running_stats
+        skills = transition.next_obs[:, -self._config.num_skills :]
+        reward = jnp.sum(
+            jax.nn.log_softmax(
+                self._discriminator.apply(discriminator_params, next_state_desc)
             )
-
-        log_q_phi = self._dynamics.apply(
-            training_state.dynamics_params,
-            state_desc,
-            active_skills,
-            target,
+            * skills,
+            axis=1,
         )
-
-        # Estimate prior skill
-        skill_samples = jnp.tile(
-            jnp.eye(self._config.num_skills), (state_desc.shape[0], 1)
-        )
-        state_descriptors = jnp.repeat(state_desc, self._config.num_skills, axis=0)
-        target = jnp.repeat(target, self._config.num_skills, axis=0)
-        log_p_s = self._dynamics.apply(
-            training_state.dynamics_params,
-            state_descriptors,
-            skill_samples,
-            target,
-        )
-        log_p_s = log_p_s.reshape((-1, self._config.num_skills))
-
-        # Compute the reward according to DADS official implementation
-        reward = jnp.log(self._config.num_skills) - jnp.log(
-            jnp.exp(jnp.clip(log_p_s - log_q_phi.reshape((-1, 1)), -50, 50)).sum(axis=1)
-        )
-
+        if add_log_p_z:
+            reward += jnp.log(self._config.num_skills)
         return reward
 
-    @partial(jax.jit, static_argnames=("self", "env", "deterministic", "evaluation"))
+    @partial(jax.jit, static_argnames=("self", "env", "deterministic"))
     def play_step_fn(
         self,
         env_state: EnvState,
-        training_state: DadsTrainingState,
-        env: Env,
+        training_state: DiaynTrainingState,
         skills: Skill,
+        env: Env,
         deterministic: bool = False,
-        evaluation: bool = False,
-    ) -> Tuple[EnvState, DadsTrainingState, QDTransition]:
+    ) -> Tuple[EnvState, DiaynTrainingState, QDTransition]:
         """Plays a step in the environment. Concatenates skills to the observation
         vector, selects an action according to SAC rule and performs the environment
         step.
@@ -258,14 +226,12 @@ class DADS(SAC):
             training_state: the DIAYN training state
             skills: the skills concatenated to the observation vector
             env: the environment
-            deterministic: the whether or not to select action in a deterministic way.
+            deterministic: whether or not to select action in a deterministic way.
                 Defaults to False.
-            evaluation: if True, collected transitions are not used to update training
-                state. Defaults to False.
 
         Returns:
             the new environment state
-            the new DADS training state
+            the new DIAYN training state
             the played transition
         """
 
@@ -292,22 +258,6 @@ class DADS(SAC):
             next_state_desc = next_env_state.info["state_descriptor"]
         else:
             next_state_desc = jnp.zeros((next_env_state.obs.shape[0], 2))
-
-        if self._config.normalize_target:
-            if self._config.descriptor_full_state:
-                _state_desc = obs[:, : -self._config.num_skills]
-                _next_state_desc = next_obs[:, : -self._config.num_skills]
-                target = _next_state_desc - _state_desc
-            else:
-                target = next_state_desc - state_desc
-
-            target *= jnp.expand_dims(1 - next_env_state.done, -1)
-            normalization_running_stats = update_running_mean_std(
-                training_state.normalization_running_stats, target
-            )
-        else:
-            normalization_running_stats = training_state.normalization_running_stats
-
         truncations = next_env_state.info["truncation"]
         transition = QDTransition(
             obs=obs,
@@ -319,15 +269,7 @@ class DADS(SAC):
             actions=actions,
             truncations=truncations,
         )
-        if not evaluation:
-            training_state = training_state.replace(
-                random_key=random_key,
-                normalization_running_stats=normalization_running_stats,
-            )
-        else:
-            training_state = training_state.replace(
-                random_key=random_key,
-            )
+        training_state = training_state.replace(random_key=random_key)
 
         return next_env_state, training_state, transition
 
@@ -341,7 +283,7 @@ class DADS(SAC):
     )
     def eval_policy_fn(
         self,
-        training_state: DadsTrainingState,
+        training_state: DiaynTrainingState,
         eval_env_first_state: EnvState,
         play_step_fn: Callable[
             [EnvState, Params, RNGKey],
@@ -354,7 +296,7 @@ class DADS(SAC):
 
 
         Args:
-            training_state: the DADS training state
+            training_state: the DIAYN training state
             eval_env_first_state: the initial state for evaluation
             play_step_fn: the play_step function used to collect the evaluation episode
             env_batch_size: the number of environments we play simultaneously
@@ -366,6 +308,7 @@ class DADS(SAC):
             state descriptors, shape: (episode_length, env_batch_size, descriptor_size)
 
         """
+
         state, training_state, transitions = generate_unroll(
             init_state=eval_env_first_state,
             training_state=training_state,
@@ -380,7 +323,7 @@ class DADS(SAC):
 
         transitions = get_first_episode(transitions)
 
-        true_returns = jnp.nansum(transitions.rewards, axis=0)
+        true_return_per_env = jnp.nansum(transitions.rewards, axis=0)
 
         reshaped_transitions = jax.tree_map(
             lambda x: x.reshape((self._config.episode_length * env_batch_size, -1)),
@@ -397,16 +340,22 @@ class DADS(SAC):
             )
 
         diversity_rewards = self._compute_diversity_reward(
-            transition=reshaped_transitions, training_state=training_state
+            transition=reshaped_transitions,
+            discriminator_params=training_state.discriminator_params,
         ).reshape((self._config.episode_length, env_batch_size))
 
         diversity_returns = jnp.nansum(diversity_rewards, axis=0)
 
-        return true_return, true_returns, diversity_returns, transitions.state_desc
+        return (
+            true_return,
+            true_return_per_env,
+            diversity_returns,
+            transitions.state_desc,
+        )
 
     @partial(jax.jit, static_argnames=("self",))
     def _compute_reward(
-        self, transition: QDTransition, training_state: DadsTrainingState
+        self, transition: QDTransition, training_state: DiaynTrainingState
     ) -> Reward:
         """Computes the reward to train the networks.
 
@@ -415,79 +364,49 @@ class DADS(SAC):
             training_state: the current training state
 
         Returns:
-            the DADS diversity reward
+            the DIAYN diversity reward
         """
         return self._compute_diversity_reward(
-            transition=transition, training_state=training_state
-        )
-
-    @partial(jax.jit, static_argnames=("self",))
-    def _update_dynamics(
-        self, operand: Tuple[DadsTrainingState, QDTransition]
-    ) -> Tuple[Params, float, optax.OptState]:
-        """Update the dynamics network, independently of other networks. Called every
-        `dynamics_update_freq` training steps.
-        """
-        training_state, transitions = operand
-
-        dynamics_loss, dynamics_gradient = jax.value_and_grad(self._dynamics_loss_fn,)(
-            training_state.dynamics_params,
-            transitions,
-        )
-
-        (dynamics_updates, dynamics_optimizer_state,) = self._dynamics_optimizer.update(
-            dynamics_gradient, training_state.dynamics_optimizer_state
-        )
-        dynamics_params = optax.apply_updates(
-            training_state.dynamics_params, dynamics_updates
-        )
-        return (
-            dynamics_params,
-            dynamics_loss,
-            dynamics_optimizer_state,
-        )
-
-    @partial(jax.jit, static_argnames=("self",))
-    def _not_update_dynamics(
-        self, operand: Tuple[DadsTrainingState, QDTransition]
-    ) -> Tuple[Params, float, optax.OptState]:
-        """Fake update of the dynamics, called every time we don't want to update
-        the dynamics while we update the other networks.
-        """
-
-        training_state, _transitions = operand
-
-        return (
-            training_state.dynamics_params,
-            jnp.nan,
-            training_state.dynamics_optimizer_state,
+            transition=transition,
+            discriminator_params=training_state.discriminator_params,
+            add_log_p_z=True,
         )
 
     @partial(jax.jit, static_argnames=("self",))
     def _update_networks(
         self,
-        training_state: DadsTrainingState,
+        training_state: DiaynTrainingState,
         transitions: QDTransition,
-    ) -> Tuple[DadsTrainingState, Metrics]:
-        """Updates the networks involved in DADS.
+    ) -> Tuple[DiaynTrainingState, Metrics]:
+        """Updates all the networks of the training state.
 
         Args:
-            training_state: the current training state of the algorithm.
-            transitions: transitions sampled from a replay buffer.
-            random_key: a random key to handle stochasticity.
+            training_state: the current training state.
+            transitions: transitions sampled from the replay buffer.
+            random_key: a random key to handle stochastic operations.
 
         Returns:
-            The updated training state and training metrics.
+            The update training state, metrics and a new random key.
         """
-
         random_key = training_state.random_key
 
-        # Update skill-dynamics
-        (dynamics_params, dynamics_loss, dynamics_optimizer_state,) = jax.lax.cond(
-            training_state.steps % self._config.dynamics_update_freq == 0,
-            self._update_dynamics,
-            self._not_update_dynamics,
-            (training_state, transitions),
+        # Compute discriminator loss and gradients
+        discriminator_loss, discriminator_gradient = jax.value_and_grad(
+            self._discriminator_loss_fn
+        )(
+            training_state.discriminator_params,
+            transitions=transitions,
+        )
+
+        # update discriminator
+        (
+            discriminator_updates,
+            discriminator_optimizer_state,
+        ) = self._discriminator_optimizer.update(
+            discriminator_gradient, training_state.discriminator_optimizer_state
+        )
+        discriminator_params = optax.apply_updates(
+            training_state.discriminator_params, discriminator_updates
         )
 
         # udpate alpha
@@ -533,7 +452,7 @@ class DADS(SAC):
         )
 
         # Create new training state
-        new_training_state = DadsTrainingState(
+        new_training_state = DiaynTrainingState(
             policy_optimizer_state=policy_optimizer_state,
             policy_params=policy_params,
             critic_optimizer_state=critic_optimizer_state,
@@ -541,21 +460,16 @@ class DADS(SAC):
             alpha_optimizer_state=alpha_optimizer_state,
             alpha_params=alpha_params,
             target_critic_params=target_critic_params,
-            dynamics_optimizer_state=dynamics_optimizer_state,
-            dynamics_params=dynamics_params,
+            discriminator_optimizer_state=discriminator_optimizer_state,
+            discriminator_params=discriminator_params,
             random_key=random_key,
-            normalization_running_stats=training_state.normalization_running_stats,
             steps=training_state.steps + 1,
         )
         metrics = {
             "actor_loss": policy_loss,
             "critic_loss": critic_loss,
-            "dynamics_loss": dynamics_loss,
+            "discriminator_loss": discriminator_loss,
             "alpha_loss": alpha_loss,
-            "alpha": alpha,
-            "training_observed_reward_mean": jnp.mean(transitions.rewards),
-            "target_mean": jnp.mean(transitions.next_state_desc),
-            "target_std": jnp.std(transitions.next_state_desc),
         }
 
         return new_training_state, metrics
@@ -563,14 +477,14 @@ class DADS(SAC):
     @partial(jax.jit, static_argnames=("self",))
     def update(
         self,
-        training_state: DadsTrainingState,
+        training_state: DiaynTrainingState,
         replay_buffer: ReplayBuffer,
-    ) -> Tuple[DadsTrainingState, ReplayBuffer, Metrics]:
+    ) -> Tuple[DiaynTrainingState, ReplayBuffer, Metrics]:
         """Performs a training step to update the policy, the critic and the
-        dynamics network parameters.
+        discriminator parameters.
 
         Args:
-            training_state: the current DADS training state
+            training_state: the current DIAYN training state
             replay_buffer: the replay buffer
 
         Returns:
@@ -578,7 +492,6 @@ class DADS(SAC):
             the replay buffer
             the training metrics
         """
-
         # Sample a batch of transitions in the buffer
         random_key = training_state.random_key
         transitions, random_key = replay_buffer.sample(
@@ -588,32 +501,17 @@ class DADS(SAC):
 
         # Optionally replace the state descriptor by the observation
         if self._config.descriptor_full_state:
-            _state_desc = transitions.obs[:, : -self._config.num_skills]
-            _next_state_desc = transitions.next_obs[:, : -self._config.num_skills]
+            state_desc = transitions.obs[:, : -self._config.num_skills]
+            next_state_desc = transitions.next_obs[:, : -self._config.num_skills]
             transitions = transitions.replace(
-                state_desc=_state_desc, next_state_desc=_next_state_desc
+                state_desc=state_desc, next_state_desc=next_state_desc
             )
 
-        # Compute the reward
-        rewards = self._compute_reward(
-            transition=transitions, training_state=training_state
-        )
+        # Compute the rewards and replace transitions
+        rewards = self._compute_reward(transitions, training_state)
+        transitions = transitions.replace(rewards=rewards)
 
-        # Compute the target and optionally normalize it for the training
-        if self._config.normalize_target:
-            next_state_desc = normalize_with_rmstd(
-                transitions.next_state_desc - transitions.state_desc,
-                training_state.normalization_running_stats,
-            )
-
-        else:
-            next_state_desc = transitions.next_state_desc - transitions.state_desc
-
-        # Update the transitions
-        transitions = transitions.replace(
-            next_state_desc=next_state_desc, rewards=rewards
-        )
-
+        # update params of networks in the training state
         new_training_state, metrics = self._update_networks(
             training_state, transitions=transitions
         )
