+++ qdax/baselines/genetic_algorithm.py
@@ -100,7 +100,7 @@ class GeneticAlgorithm:
             emitter_state: state of the emitter
             random_key: a jax PRNG random key
 
-        Results:
+        Returns:
             the updated MAP-Elites repertoire
             the updated (if needed) emitter state
             metrics about the updated repertoire
