+++ tests/core_test/diayn_test.py
@@ -9,7 +9,7 @@ import pytest
 from brax.envs import State as EnvState
 
 from qdax import environments
-from qdax.core.diayn import DIAYN, DiaynConfig, DiaynTrainingState
+from qdax.baselines.diayn import DIAYN, DiaynConfig, DiaynTrainingState
 from qdax.core.neuroevolution.buffers.buffer import QDTransition, ReplayBuffer
 from qdax.core.neuroevolution.sac_utils import do_iteration_fn, warmstart_buffer
