+++ tests/core_test/td3_test.py
@@ -6,9 +6,9 @@ import pytest
 from brax.envs import State as EnvState
 
 from qdax import environments
+from qdax.baselines.td3 import TD3, TD3Config, TD3TrainingState
 from qdax.core.neuroevolution.buffers.buffer import ReplayBuffer, Transition
 from qdax.core.neuroevolution.mdp_utils import do_iteration_fn, warmstart_buffer
-from qdax.core.td3 import TD3, TD3Config, TD3TrainingState
 
 
 def test_td3() -> None:
